feat(redis): allow custom redis url
This new feature removes the old `REDIS_HOST` env variable and replaces it by `REDIS_URL`. If it is unset, a default redis URL is used. Fixes #22

Although the likelyhood of a breaking change is low, we introduce it as new feature as some people may have used a custom redis hostname already.